fix release pipeline + update issue templates (#98)
* fix release pipeline + update issue templates

* improve timechart screenshots

* fix ts error

* update test

* update test

* re-run tests

+++ e2e/tests/svgScreenshots/timechart.svgif.ts
@@ -2,7 +2,10 @@ import { getCommandElemSelector, getDataKey } from "Testing";
 import {
   closeWorkspaceWindows,
   deleteAllWorkspaces,
+  newChat,
   openTable,
+  sendAskLLMMessage,
+  setPromptByText,
 } from "utils/utils";
 import type { OnBeforeScreenshot } from "./SVG_SCREENSHOT_DETAILS";
 
@@ -12,9 +15,66 @@ export const timechartSvgif: OnBeforeScreenshot = async (
   { addScene, addSceneAnimation },
 ) => {
   await openConnection("crypto");
-  await deleteAllWorkspaces(page);
-  await closeWorkspaceWindows(page);
-  await openTable(page, "futures");
+
+  const wspText = await page
+    .locator(
+      getCommandElemSelector("SilverGridChild") +
+        `[data-view-type="timechart"]`,
+    )
+    .textContent();
+  const alreadyShowing =
+    wspText?.includes("Multi-Asset Price Comparison") ?? false;
+  if (!alreadyShowing) {
+    await deleteAllWorkspaces(page);
+    await closeWorkspaceWindows(page);
+    await page.getByTestId("AskLLM").click();
+    await setPromptByText(page, "dashboard");
+    await newChat(page);
+    await sendAskLLMMessage(page, " funding ");
+    await page.getByTestId("AskLLMChat.LoadSuggestedDashboards").click();
+  }
+  await page.waitForTimeout(2000);
+  const shortWait = { animations: [{ type: "wait" as const, duration: 1500 }] };
+  const chart = page
+    .locator(`${getCommandElemSelector("W_TimeChart")} canvas`)
+    .first();
+  await chart.waitFor({ state: "visible" });
+  const chartBBox = await chart.boundingBox();
+
+  /** Zoom in the center of chart */
+  const zoomSteps = 9;
+  for (let i = 0; i < zoomSteps; i++) {
+    const isLast = i === zoomSteps - 1;
+    await chart.hover();
+    await page.mouse.move((chartBBox!.x + chartBBox!.width) / 2, 300);
+    await page.waitForTimeout(100);
+    await page.mouse.wheel(0, -200);
+    // await page.waitForTimeout(100);
+    // await page.mouse.move(0, 300);
+    await page.waitForTimeout(1000);
+    await addScene({
+      animations: [
+        isLast ?
+          { type: "wait" as const, duration: 3000 }
+        : shortWait.animations[0],
+        // {
+        //   type: "custom",
+        //   elementSelector:
+        //     getCommandElemSelector("W_TimeChart") + " g g g path",
+        //   attributes: {
+        //     transform: ["scale(1, 1)", "scale(1.5, 1)"],
+        //   },
+        //   fixedAttributeValues: {
+        //     "transform-origin": "center center",
+        //   },
+        //   duration: 1500,
+        // },
+      ],
+    });
+  }
+
+  return;
+  // await openTable(page, "futures");
   await page.getByTestId("dashboard.window.toggleFilterBar").click();
   await page.getByTestId("SearchList.Input").fill("btcu");
   // await addSceneAnimation(getCommandElemSelector("SearchList.Input"), {
